+++ CHANGELOG.md
@@ -8,7 +8,9 @@
 So, you also have to import function only through default export and only with name `cn`.
 Now you have to set option `fn` with name of your function, no matter where from you've got it.
 For example, now you can set option like `fn: 'clsx'` and use named import like `import { clsx } from 'clsx'`.
-- restore removed `jsx-a11y` aribnb config 
+- restore removed `jsx-a11y` aribnb config
+- add rule `@dartess/mobx-no-flow`: Disallow the use of MobX `flow` in favor of `async`/`await` with actions
+- add `@dartess/mobx-no-flow` to `mobx` config
 
 ## [0.12.0] - 2026-04-22